Due to mono port, license can be updated to be more clear.
The original agreement was to distribute the code as GPL 3.0 or later and only use LGPL 3.0 or later if necessary. At the momment, the only potential issue is any restrictions placed by the Windows version, which is compiling and run using DOT NET 4.5+. I'm not sure if this is necessary, but to avoid any potential licensing issues in the future, LGPL 3.0 should only apply if necessary for the windows version. Mac/Linux/Unix-like should never require LGPL if mono remains free.